Move updatable notification types to /models
