+++ src/client.spec.ts
@@ -33,7 +33,6 @@ import {
 	SystemStatusCategory,
 	SystemStatusMessageReceived,
 	SystemStatusMessageSent,
-	UpdatableNotification,
 	UpdatablePlaybook,
 	UpdatableResource,
 	UpdatableSavedQuery,
@@ -88,6 +87,7 @@ import {
 	UpdatableFile,
 	UpdatableGroup,
 	UpdatableMacro,
+	UpdatableNotification,
 	UploadableAutoExtractor,
 	User,
 	UserPreferences,
