+++ src/functions/notifications/update-one-notification.ts
@@ -6,14 +6,13 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { toRawNumericID } from '../../value-objects';
+import { toRawUpdatableNotification, UpdatableNotification } from '../../models/notification';
 import {
 	APIFunctionMakerOptions,
 	buildHTTPRequest,
 	buildURL,
 	fetch,
 	HTTPRequestOptions,
-	omitUndefinedShallow,
 	parseJSONResponse,
 } from '../utils';
 
@@ -29,7 +28,7 @@ export const makeUpdateOneNotification = (makerOptions: APIFunctionMakerOptions)
 
 			const baseRequestOptions: HTTPRequestOptions = {
 				headers: { Authorization: authToken ? `Bearer ${authToken}` : undefined },
-				body: JSON.stringify(toUpdatableRaw(updatable)),
+				body: JSON.stringify(toRawUpdatableNotification(updatable)),
 			};
 			const req = buildHTTPRequest(baseRequestOptions);
 
@@ -41,31 +40,3 @@ export const makeUpdateOneNotification = (makerOptions: APIFunctionMakerOptions)
 		}
 	};
 };
-
-export interface UpdatableNotification {
-	id: string;
-	message?: string;
-	customID?: string;
-
-	sentDate?: string; // Timestamp eg. '2019-04-22T21:44:01.776942432Z'
-	expirationDate?: string; // Timestamp eg. '2019-04-23T03:44:01.776918756-06:00'
-	ignoreUntilDate?: string; // Timestamp eg. '0001-01-01T00:00:00Z'
-}
-
-interface RawUpdatableNotification {
-	Msg?: string;
-	Type?: number;
-	Sent?: string; // Timestamp eg. '2019-04-22T21:44:01.776942432Z'
-	Expires?: string; // Timestamp eg. '2019-04-23T03:44:01.776918756-06:00'
-	IgnoreUntil?: string; // Timestamp eg. '0001-01-01T00:00:00Z'
-}
-
-const toUpdatableRaw = (updatable: UpdatableNotification): RawUpdatableNotification =>
-	omitUndefinedShallow({
-		Msg: updatable.message,
-		Type: updatable.customID === undefined ? undefined : toRawNumericID(updatable.customID),
-
-		Sent: updatable.sentDate,
-		Expires: updatable.expirationDate,
-		IgnoreUntil: updatable.ignoreUntilDate,
-	});
