+++ src/models/notification/index.ts
@@ -16,6 +16,9 @@ export * from './raw-creatable-broadcasted-notification';
 export * from './raw-creatable-targeted-notification';
 export * from './raw-notification';
 export * from './raw-targeted-notification';
+export * from './raw-updatable-notification';
 export * from './targeted-notification';
 export * from './to-raw-creatable-broadcasted-notification';
 export * from './to-raw-creatable-targeted-notification';
+export * from './to-raw-updatable-notification';
+export * from './updatable-notification';
